Serializer/Importer: Move Content and LargeContent into RT::Record
The logic previously resided on both RT::Attachment and
RT::ObjectCustomFieldValue to ensure that the raw bytes (and not the
encoding as they exist in the DB) are written to the serialized stream,
and that they are writen in the correct encoding on the remote end.
Combine them on RT::Record for greater generality.

This most probably fixes a bug wherein ObjectCustomFieldValue's lack of
Serialize method caused it to possibly insert _encoded_ information.