Serializer/Importer: Export and respect clone property in stream metadata
Doing so requires that we delay some initialization that was previously
done at new()-time until when the metadata is received, which may not
happen until we have called ->Import to open the stream.  Move these
aspects of initialization to a new ->InitStream method, which is called
once this information is available.