Clean out any lingering joins set in _Init
Users and Groups automatically join to Principals.  With an old
enough version of DBIx::SearchBuilder (3.2 era, v1.27 of DBIx::SB)
you end up with a cartesian product query.  The old SB doesn't generate
a JOIN, it generates a WHERE (main.id = Principals_1.id) and
unfortunately, our NextPage code uses ENTRYAGGREGATOR = 'none' to build
a main.id > $old_id limit and ends up losing the join clause.  We want
that NextPage code because it allows incremental exports and clones of
large datasets to finish in a reasonable amount of time, so this works
around the old SB bugs.  It also avoids any future surprises in _Init.